resolve transformers relative to config file
For consistency, to match user expectations, and to make dependency
handling sane, it makes more sense to resolve transformer modules
relative to the configuration file in which they are declared.
This allows invoking openapi-transformer-cli from anywhere, with a local
or global installation, and getting consistent and reasonable results,
including using dependencies declared in the project which includes the
transformer config file.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>